+++ src/main/java/io/github/whatss7/onlyflax/OnlyFlax.java
@@ -7,25 +7,21 @@ import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.ModLoadingContext;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.fml.config.ModConfig;
-import net.neoforged.fml.javafmlmod.FMLJavaModLoadingContext;
 
 @Mod(OnlyFlax.MOD_ID)
 public class OnlyFlax {
     public static final String MOD_ID = "onlyflax";
 
-    public OnlyFlax() {
+    public OnlyFlax(IEventBus modBus) {
         // Load config
         ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, OnlyFlaxConfig.COMMON_SPEC);
 
-        // Get eventBus
-        IEventBus eventBus = FMLJavaModLoadingContext.get().getModEventBus();
-
         // Register items & blocks
-        ModBlocks.register(eventBus);
-        ModItems.register(eventBus);
-        ModItems.setupItemUsages(eventBus);
+        ModBlocks.register(modBus);
+        ModItems.register(modBus);
+        ModItems.setupItemUsages(modBus);
 
         // Register wild flax features
-        ModBiomeModifiers.register(eventBus);
+        ModBiomeModifiers.register(modBus);
     }
 }