+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 1.0.3
 
-- Ported to NeoForge 1.20.2 from 1.20.1-forge-1.0.3
+- Ported to NeoForge 1.20.4 from 1.20.1-forge-1.0.3